# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta es la API de RPL 2.0 para todos y todas",
        "version": "0.1",
        "title": "RPL 2.0",
        "termsOfService": "Terms of service",
        "contact": {
            "name": "Alejandro Levinas o Mat\u00edas Cano",
            "url": "http://example.com/support",
            "email": "levinasale@gmail.com;matiasjosecc@gmail.com"
        },
        "license": {
            "name": "License of API",
            "url": "http://example.com/support"
        }
    },
  "host": "localhost:38497",
    "basePath": "/",
...
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "authentication-controller"
                ],
                "summary": "authenticateUser",
                "operationId": "authenticateUserUsingPOST",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "loginRequestDto",
                        "description": "loginRequestDto",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/LoginRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/JwtResponseDTO"
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/api/auth/profile": {
            "get": {
                "tags": [
                    "authentication-controller"
                ],
                "summary": "getUser",
                "operationId": "getUserUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "accountNonExpired",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "accountNonLocked",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "authorities[0].authority",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "credentialsNonExpired",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "enabled",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.dateCreated",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "format": "date-time"
                    },
                    {
                        "name": "user.degree",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.email",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.emailValidated",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "user.id",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "user.lastUpdated",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "format": "date-time"
                    },
                    {
                        "name": "user.name",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.password",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.studentId",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.surname",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.university",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.username",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserResponseDTO"
                        }
                    },
                    "403": {
                        "description": "Forbidden!!!!!"
                    },
                    "500": {
                        "description": "500 message",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/api/auth/signup": {
            "post": {
                "tags": [
                    "authentication-controller"
                ],
                "summary": "registerUser",
                "operationId": "registerUserUsingPOST",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "createUserRequestDTO",
                        "description": "createUserRequestDTO",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateUserRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserResponseDTO"
                        }
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "definitions": {
...
        "CreateUserRequestDTO": {
            "type": "object",
            "properties": {
                "degree": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "student_id": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "university": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            },
            "title": "CreateUserRequestDTO"
        },
...
        "JwtResponseDTO": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "token_type": {
                    "type": "string"
                }
            },
            "title": "JwtResponseDTO"
        },
        "LoginRequestDTO": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username_or_email": {
                    "type": "string"
                }
            },
            "title": "LoginRequestDTO"
        },
...
        "UserResponseDTO": {
            "type": "object",
            "properties": {
                "date_created": {
                    "type": "string",
                    "format": "date-time"
                },
                "degree": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "email_validated": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "last_updated": {
                    "type": "string",
                    "format": "date-time"
                },
                "name": {
                    "type": "string"
                },
                "student_id": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "university": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            },
            "title": "UserResponseDTO"
        }
    }
}